+++ renderer/src/data/changelog.js
@@ -1,21 +1,24 @@
 export default {
-    description: "This version is a hotfix mainly to fix the settings issue.",
+    description: "Most of these fixes in this update come from Strencher (https://github.com/Strencher).",
     changes: [
         {
             title: "Fixes",
             type: "fixed",
             items: [
-                "Fixed an issue with the settings patch failing",
-                "Fixed an issue with hiding GIF and Gift buttons",
-                "Fixed certain plugin patches not returning values",
-                "Updated links to new website"
+                "Fixed several crashing issues related to voice and video chats.",
+                "Fixed the incompatibility with Canary/PTB",
+                "Fixed missing class names on guild elements.",
+                "Fixed minimal mode for new Discovery tab. Thanks, dav1312 (https://github.com/dav1312)."
             ]
         },
         {
             title: "Power Users & Developers",
             type: "improved",
             items: [
-                "Certain patches done using `monkeyPatch` instead of `Patcher` now properly return values."
+                "Console errors from plugins should now properly map to the local file.",
+                "Support for other chrome profiles added when using React DevTools. Thanks, CrizGames (https://github.com/CrizGames).",
+                "Emote menu handling is improved.",
+                "`openDialog` added to `BdApi`."
             ]
         }
     ]