+++ renderer/src/modules/core.js
@@ -104,7 +104,7 @@ export default new class Core {
     }
 
     async checkForUpdate() {
-        const resp = await fetch(`https://api.github.com/repos/rauenzi/BetterDiscordApp/releases/latest`,{
+        const resp = await fetch(`https://api.github.com/repos/BetterDiscord/BetterDiscord/releases/latest`,{
             method: "GET",
             headers: {
                 "Accept": "application/json",